refactor hotkey components
