+++ src/renderer/src/components/settings/hotkeys/SavingIndicator.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import styles from '../HotkeyManager.module.css';
+
+interface SavingIndicatorProps {
+  saving: boolean;
+}
+
+export const SavingIndicator: React.FC<SavingIndicatorProps> = ({ saving }) => {
+  if (!saving) {
+    return null;
+  }
+
+  return <div className={styles.savingIndicator}>Saving settings...</div>;
+};
