+++ src/renderer/src/components/settings/hotkeys/index.ts
@@ -1,0 +1,7 @@
+export { useHotkeyManager } from './useHotkeyManager';
+export { LoadingState } from './LoadingState';
+export { HotkeyHeader } from './HotkeyHeader';
+export { GlobalToggle } from './GlobalToggle';
+export { HotkeyList } from './HotkeyList';
+export { HotkeyInstructions } from './HotkeyInstructions';
+export { SavingIndicator } from './SavingIndicator';
